🎨: デザインシステム連携のため上に戻るリンクパターンの記述構成を見直す
