+++ src/patterns/UpwardNavigation/UpwardNavigation.stories.tsx
@@ -1,34 +1,13 @@
-import styled, { css } from 'styled-components'
 import { Story } from '@storybook/react'
-import { FaArrowLeftIcon, Heading, Stack, Text, TextLink } from 'smarthr-ui'
+import { ComponentProps } from 'react'
 
-// TODO 共通のやつは置き方や使い方を考えたい
-import { GlobalNav } from '../../GlobalNav/GlobalNav'
-import { UpwardLinkWrapper } from '@patterns/UpwardNavigation'
+import { GlobalNav } from '@GlobalNav/GlobalNav'
+import { UpwardNavigation } from './UpwardNavigation'
 
-type StoryProps = {
-  withAppNavi: boolean
-  withUpwardLink: boolean
-}
-
-const Template: Story<StoryProps> = ({ withAppNavi, withUpwardLink }) => (
+const Template: Story<ComponentProps<typeof UpwardNavigation>> = ({ withAppNavi, ...props }) => (
   <>
     <GlobalNav withAppNavi={withAppNavi} />
-    <Wrapper>
-      {withUpwardLink && (
-        <UpwardLinkWrapper withAppNavi={withAppNavi}>
-          <TextLink href="#" prefix={<FaArrowLeftIcon />}>
-            分析レポートに戻る
-          </TextLink>
-        </UpwardLinkWrapper>
-      )}
-      <Stack>
-        <Heading>分析対象の従業員項目</Heading>
-        <Text as="p">
-          一部の数値データにおいて、クロス集計する際の集計単位を変更できます例えば「年齢」を「60」以上をまとめる、「20」以下をまとめる、「10」単位でまとめるのように設定すると、以下のように出力されます。
-        </Text>
-      </Stack>
-    </Wrapper>
+    <UpwardNavigation {...props} withAppNavi={withAppNavi} />
   </>
 )
 
@@ -52,14 +31,6 @@ NoUpwardNav.args = {
   withUpwardLink: false,
 }
 NoUpwardNav.storyName = '上に戻るリンクなし'
-
-const Wrapper = styled(Stack).attrs({ gap: 1.5 })`
-  ${({ theme: { space } }) => css`
-    margin-inline: auto;
-    padding: ${space(2)} ${space(1.5)};
-    max-width: 80em;
-  `}
-`
 
 export default {
   title: '上に戻るリンク',
