+++ src/patterns/UpwardNavigation/index.ts
@@ -1,1 +1,1 @@
-export { UpwardLinkWrapper } from './UpwardLinkWrapper'
+export { UpwardLinkWrapper } from './UpwardNavigation'
